requirements.txt
streamlit
google-generativeai
transformers
torch
scikit-learn
pandas
